+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 ## CHANGE LOG ##
 
+* v2.2.0 (2017-02-  )
+   * [RB-5] Fixed error code range not being checked when used with custom message strings
+   * For sake of logic, errorWithDataAndHttpCode() and errorWithHttpCode() will throw exception if http_code is null
+   * http_code can be handed as null to all the other methods to have it replaced by default code
+
 * v2.1.2 (2016-08-24)
    * Fixed exception code handling in ExceptionHandlerHelper (reported by Adrian Chen @absszero)